learnlive/requirements.txt
Django==1.5.1
MySQL-python==1.2.5
PyYAML==3.10
South==0.8.4
distribute==0.7.3
django-appconf==0.6
django-compressor==1.3
django-debug-toolbar==1.0.1
django-pipeline==1.3.20
django-treebeard==1.61
futures==2.1.6
httplib2==0.8
nltk==2.0.4
numpy==1.8.0
opentok==0.91.2
pycrypto==2.6.1
pydot==1.0.28
pyparsing==2.0.1
python-graph-core==1.8.2
python-graph-dot==1.8.2
six==1.5.2
sqlparse==0.1.11
textblob==0.8.0
twilio==3.6.5
wsgiref==0.1.2
